refactor: context api 사용처 리팩토링
- useMemo, useCallback 적용
- 폴더 구조 변경
  - context
  ㄴ ...Context
  ㄴ ...Provider
